+++ src/main/java/com/jaquadro/minecraft/storagedrawers/inventory/ContainerDrawersComp3.java
@@ -5,18 +5,18 @@ import com.jaquadro.minecraft.storagedrawers.core.ModContainers;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
 
-public class ContainerDrawersComp extends ContainerDrawers
+public class ContainerDrawersComp3 extends ContainerDrawers
 {
     private static final int[][] slotCoordinates = new int[][] {
         { 80, 23 }, { 67, 49 }, { 93, 49 }
     };
 
-    public ContainerDrawersComp (int windowId, Inventory playerInventory, FriendlyByteBuf packet) {
-        super(ModContainers.DRAWER_CONTAINER_COMP.get(), windowId, playerInventory, packet);
+    public ContainerDrawersComp3 (int windowId, Inventory playerInventory, FriendlyByteBuf packet) {
+        super(ModContainers.DRAWER_CONTAINER_COMP_3.get(), windowId, playerInventory, packet);
     }
 
-    public ContainerDrawersComp (int windowId, Inventory playerInventory, BlockEntityDrawers tile) {
-        super(ModContainers.DRAWER_CONTAINER_COMP.get(), windowId, playerInventory, tile);
+    public ContainerDrawersComp3 (int windowId, Inventory playerInventory, BlockEntityDrawers tile) {
+        super(ModContainers.DRAWER_CONTAINER_COMP_3.get(), windowId, playerInventory, tile);
     }
 
     @Override