+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[12.7.0]
+- Added 2-tier version of compacting drawers
+- Added half-depth versions of both compacting drawers
+- Max capacity checking for pulled drawers is default off in new configs
+- BREAKING: Some resource filenames related to compacting drawers have changed, which may affect data packs.  Compacting drawers will likely fall back to their default resources.
+
 [12.6.6]
 - Fixed pulled drawer losing its capacity data if pulled when empty